addItem도 nav 활성화 수정
